Disable auto-download and hard-coded checksums, use .sha256 files
Summary:
This diff simplifies automated building of NPMs.

It drops hard-coded checksums in favor of .sha256 files (generated by
sha256sum) since the purpose is to ensure complete and correct downloads, and
not to verify provenance. This removes the need for manually updating them.

It also drops support for downloading releases, which simplifies the release
process by not requiring a GitHub release before an NPM release.

Reviewed By: mhorowitz

Differential Revision: D16604135

fbshipit-source-id: 45dcb491fd0910d5448cdf517733d0045ae27ab5