Use VTable functions for custom snapshot output
Summary:
Some objects are more important to the user than others, and might want to heavily customize what
they look like in the snapshot output.

The most obvious case is `JSObject`, which is what all user types are currently classified as.
Objects might also want to show properties (named things like `o.foo`) and elements (`o[4]`) in
different ways. They may also have internal pointers that are important to the VM authors such as
their `HiddenClass`. They also want to have a custom way of finding a type name, since they could
be ES6 class objects or small object literals.

The main thing to notice about this API is that `addEdges` is called twice for every object.
* The first call counts the number of edges
* The second call outputs the edges

I felt it was best to only have one function on `VTable` adds edges, rather than an additional one
that counts the number of edges.
The unfortunate side effect of this is that if discovering edges is expensive, it will be done twice
needlessly; however, taking a snapshot is not a time-critical event, so some duplicated work is
tolerable.

Reviewed By: davedets

Differential Revision: D16472216

fbshipit-source-id: fad88abca94ae1fad24f1f3653c74ecb5b20db56