Handle relocation of native function pointers.
Summary:
[Serialize/Deserialize NativeFunction/NativeConstructor] Serialize/Deserialize needs a way to relocate native
function pointers. Map the pointer to an id when serializer is
initialized and use this id as the relocation id for the function pointer.
This mapping order is defined by NativeFunctions.def.

This commit handles the relocation part. The Serialzer and the Deserializer
reads the same .def file therefore can have the same id for the same function.

Reviewed By: dulinriley

Differential Revision: D16271413

fbshipit-source-id: 984438234b66e765c8a49a6b0065ff6b592c219e