New HERMES_EXTRA_DEBUG code to try to debug card-object-table problems.
Summary: There's a crash happening in the wild that we're not able to reproduce locally.  It manifests in young-gen card table scanning: a dirty card leads to a first spanning object that is not valid.  I have some reason to believe the card table boundary calculation leads us outside the segment (from examining a minidump).  But the minidump information is partial, and inconclusive.  This diff adds extra debug information to Android custom breakpad data when this bug is about to happen, which I hope will advance us towards a diagnosis.

Reviewed By: tmikov

Differential Revision: D16754036

fbshipit-source-id: 03a539e0eaf8d534e5ccea8c259af01d0af8a08b