Infer type name for objects in heap snapshot
Summary:
Most user-created classes that look like this:
```
function MyObject(x) {
  this.x = x;
}
```
would prefer to see the type as `MyObject` in the heap instead of `Object`.
This requires using the `obj.constructor.name` if it exists. This requires carefully
looking for the property in the prototype chain without doing any allocations in
`JSObject::tryGetNamedFast`.

If it doesn't exist, use the `CellKind` as a backup.
In the future, more estimates can be added; for example,
v8 uses the name of the local variable the object is assigned to if it can
so that `var foo = {bar: 1, baz: {}};` gets reported as a `foo`.

Reviewed By: amnn

Differential Revision: D16508749

fbshipit-source-id: fa9edd6eef42110635f80b17371587d93a901fd1