Remove the @ prefix for internal pointers in snapshots
Summary:
There used to be a worry that property names in user-made properties would conflict with
internal properties.
But Chrome heap snapshots treat the combination of `(EdgeType, name_or_index)` as unique,
and only internal properties have the `EdgeType::Internal`. The Chrome visualizer also makes
it clear the properties are internal by graying them out.

So might as well remove the unnecessary prefix to make it easier to read internal properties.

Reviewed By: jbower-fb

Differential Revision: D16772482

fbshipit-source-id: 854bfe44df4840ef71d6e934fb8a423ef86ac5c0