Remove non pointer fields from metadata
Summary:
There's no way for Chrome heap snapshots to display inline values such as numbers, so remove
them from Metadata.

If the Chrome format is extended later, we can add customizations in `addEdgesImpl` as well, no need
to put them in the Metadata.
This helps keep the metadata compact in memory and easier to scan by the GC.

Reviewed By: jbower-fb

Differential Revision: D16772483

fbshipit-source-id: 8b9e4b102bc6c4a3e7998ce9bc015736418f95c7