Skip internal properties in JSObject
Summary:
Internal properties are sometimes meant to only be internal the implementation of the VM,
and not a user-visible property that can be mutated.
To avoid confusing the user, avoid emitting any name for these properties.

There's a separate question to ask here: should we name these something other than `directPropX` if
we have an internal name? Maybe, but it would be complicated and not necessary for now.
Simply ignoring the user-visible version is much easier and doesn't lose any information in the snapshot.

Reviewed By: mhorowitz

Differential Revision: D16777340

fbshipit-source-id: 4009ff344d44ffd78581f939ca23e21bb74893a1