Use RootSections for heap snapshots
Summary:
In order to better classify what keeps values alive, it is beneficial to have the roots broken down
by section. Hermes's sections are things like the registers, identifier table, custom roots, and so on.

To do this, add `beginRootSection` and `endRootSection` to `SlotAcceptor`. It defaults to an empty
implementation for most acceptors, but snapshots use it to construct new nodes and edges to those
new nodes.

Reviewed By: amnn

Differential Revision: D16767262

fbshipit-source-id: aa598c83883ac950197d38555370d444a42b1323