Remove HERMESVM_API_TRACE from the VM
Summary:
Now that unique object IDs are available in all build modes, it's time to make synthetic benchmarks
available all the time!

The first step is removing the `ifdefs` inside the VM so that `MockedEnvironment` can be used always.
This involves an extra branch in the following functions:
* `Date.now()`
* `new Date(...)`
* `Date(...)`
* `Math.random()`

These functions are not critical for performance, and the branch should be well-predicted.
Note that this diff only covers tracing. Replay will be handled in a separate commit.

Reviewed By: jbower-fb

Differential Revision: D16808268

fbshipit-source-id: a7fa138a0c7d5e1028906af78e117935332cea50