Serialize/deserialize ArrayStorage with its owner.
Summary:
Serializer and Deserialzer cannot distinguish NativePointer HermesValue and
NativeValue HermesValue, they have to rely on the entity that calls them to
tell them whether a HermesValue is a NativePointer or not. This works most of
the cases except for ArrayStorage. ArrayStorages used by Domain store some
NativePointers while other ArrayStorages don't. In order to
serialize/deserialzie ArrayStorage correctly, we need to be able to know what
kind of ArrayStorage it is.

Because every ArrayStorage has exactly one owner, instead of serialize/deserialize
ArrayStorage by themselves when we see them on the Heap, we can
serailize/deserialize ArrayStorage with its owner (i.e. we serialize
ArrayStorage GCCell when we serialize the GCCell that creates the
ArrayStorage). Therefore, we know if the ArrayStorage has any native pointers
when we serialize/deserialize it.

This also means we will skip ArrayStorage GCCells when we sweep the heap.

Reviewed By: avp

Differential Revision: D16623982

fbshipit-source-id: 9db9ecb1300042c9d3fb06a9d61dac09b2fffb89