Serialize/Deserialize RuntimeModule, BCProvider, CodeBlock.
Summary:
Straightforward implementation of serializing/deserializing
RuntimeModule, BCProvider, CodeBlock.

General principle of serialization/deserialization with Domain, RuntimeModule,
BCProvider, and CodeBlock is that an entity should be serialized/deserialized
with its owner. We define "owns" the same way as we handle ownership for memory
management at runtime (as described in Design.md). That is: Domain owns a list
of RuntimeModules, and RuntimeModule owns a list of CodeBlocks.

We also need to decide what to serialize/deserialize for BCProvider and
CodeBlocks. For a straightforward first design, we will not support
serialize/deserialize lazy functions. We serialize
BCProviderFromBuffer's buffer directly and use ByteCodeSerializer to
serialize BCProviderFromSrc's ByteCodeModule in Bytecode file format.
Therefore, when we are desreializing, we just read the buffer as a Bytecode
file regardless.

For CodeBlock, its entity is its function id, meaning that as long as
bcProvider_ is materialized before deserializing a CodeBlock, we can
reconstruct the CodeBlock with its id.

Reviewed By: avp

Differential Revision: D16618633

fbshipit-source-id: ee3cde58bf75f2624f0ddf4c743143fb8e442e30