Add property serialization to Object names in snapshots
Summary:
For objects not in dictionary mode, include their property names as part of the name
of the type.
This works well to group objects related to similar systems (all module exports are categorized this way).

However, it does have a problem of "over-categorization", there are type names with only 1 or 2 objects.
It would be nice to avoid giving those objects their own category, but doing that would require a post-process step
after all object names were determined (in order to determine how many objects are in each category).

Reviewed By: amnn

Differential Revision: D17183827

fbshipit-source-id: 81e66b2852f38cca860862eedb9a168f65f9aea4