Fix displayed value of floating point constants in -dump-bytecode mode
Summary:
In -dump-bytecode mode, the displayed values of floating point constants are incorrect
eg. 4.614253e+18 is displayed when the correct value is 3.14

It was because the `decodeOperand` function treats the return value as int64 regardless of its actual type.
Modified `decodeOperand` to cast return value to double when the operand is a double.

Reviewed By: dulinriley

Differential Revision: D17213601

fbshipit-source-id: a9f6617179b8d992c4f7d722ce058775f278dc7f