mmap the register stack to reduce resident size
Summary:
When the VM allocates its own register stack, use mmap rather than malloc. We typically use a very small fraction of the stack reservation. But depending on the malloc implementation, the pages may still be dirty from the OS point of view. With mmap, only the pages actually used will be resident.

In the future, we could explicitly release unused pages after the stack has shrunk (and we could do it as part of full GC), but that's not implemented in this diff.

Reviewed By: dulinriley

Differential Revision: D17209165

fbshipit-source-id: 3669ad2c3519cbe5fb6940bbdedfe773eb5d272c