Handle error in fb4a-marketplace-global-code benchmark
Summary:
`fb4a-marketplace-global-code` is a synthetic benchmark in gc-bench, and it was failing with exception thrown from `hermes/API/hermes/TraceInterpreter.cpp`. The problem was with `markerFound` was only set to true for `RecordType::Marker` and not for  `RecordType::EndExecJS`.

This commit makes `case RecordType::EndExecJS` fall through to `Record::Marker` since `RecordType::EndExecJS` is a subclass of `Record::Marker`.

Reviewed By: dulinriley

Differential Revision: D17287970

fbshipit-source-id: 75f9e0755e2f282ec9800d90917c074c43f9fd82