Grow IdentifierTable's vector less aggressively
Summary:
Change IdentifierTable's lookup vector to expand its capacity less aggressively (1.25x instead of the usual 2x).

This wastes less space in the common case where the number of identifiers created dynamically (D) is small compared to the number of identifiers initialized from the (largest) module (M). With an ordinary vector, since we reserve M upfront, the addition of D would result in a 2*M capacity.

Reviewed By: ridiculousfish

Differential Revision: D17352970

fbshipit-source-id: fee1ba42e7a1428b17bfcce19e371cdfbfd7c577