Squash diffs for: A fix to card boundary table protection; and Set a max page size, align segment metadata components to that size.
Summary:
This diff is the squash of two approved diffs.

Squashing because the first without the second would cause platforms with 16K pages to crash.  Would be landing as a stack, but I didn't want that to potentially be a problem with, say, bisection searches.

The diff descriptions:

----------------------
In the debacle that followed from my previous attempt to do card boundary table protection, we discovered that we don't have test coverage for the GC mode actually being used by RN apps: "no GC before TTI."  This diff adds three tests of that mode.  Each of them runs the deltablue benchmark in "no-gc-before-TTI" mode.
We run some iterations, then declare TTI, before needing to do a full GC.  Then we run more iterations.  This exposed a bug observed when running (e.g.) Marketplace: in debug builds, a path that does allocSlow would protect the active segment, and then HermesInternal.ttiReached would call OldGen::recreateCardBoundaryTable, which would segv because of the protection.
We run enough iterations to reach a full GC before declaring TTI.
We do an explicit GC before declaring TTI.  This exposed a bug in the VM protection: we were iterating over segments before yielding them.
In addition to adding these tests, this diff fixes the bugs they expose:
OldGen::recreateCardBoundaryTable unprotects and reprotects the boundary table.
GenGC::collect moves it's unprotection call after yielding the allocation context.
It also generally makes the mechanism more robust:
It adds a tracking mechanism to make sure that boundary tables are protected when we unprotect them, and unprotected when we protect them.
We take care to protect boundary tables of all segments: the initial active segment is protected in the OldGen ctor, and we protect tables when we materializeNewSegments outside of GC, as a result of direct OG allocation.
------------------------
We introduced card table boundary table memory protection.  For this to work properly on all platforms, we need the metadata component that's being protected to be page-aligned and have a size that's a multiple of the page size.  This diff makes that true.
First, we introduce a constant in a new file, ExpectedPageSize.h:  pagesize::kExpectedPageSize.  The intent is that this is at least as large as the actual page size on the platform; using this for alignment and size will guarantee page alignment and size.  We use round up the size of the card table, card boundary table, and mark bit array to this size, thus ensuring that all the boundaries between these, and between the mark bits and the alloc region, will have kMaxPageSize alignment.  Thus, we can successfully mprotect the card boundary table (and, in the future, the mark bit array, if we want).
We adopt a suggestion from Daniel and Tzvetan: in case the static constant is wrong for some platform, we check at runtime whether the actual page size is in fact <= the kExpectedPageSize.  If not, we disable the mprotect mechanism.  (We put the instance field controlling this in GenGC, because it might eventually apply to mark bit arrays, which are used in both old- and young-generation segments.)

Reviewed By: tmikov

Differential Revision: D17408450

fbshipit-source-id: 6f1fc820742801f29e4fa2029b42779e312f9014