Fix handling of notes in SourceErrorManager
Summary:
Notes were being treated as independent errors, which produced
undesirable results during sorting or when a suppressed warning had
notes. So, instead we associate the note with the previous message. If
that message is suppressed, suppress the note too. When sorting, sort
the notes together with the previous non-note message.

Reviewed By: avp

Differential Revision: D17485317

fbshipit-source-id: 2e9643471f55d883acd8a221a9c068f6b2d9d5e5