Expose RuntimeConfig default values
Summary:
Add the ability to get the default value of a RuntimeConfig parameter, also as constexpr (for the types that support it).

This can help to keep defaults synchronized between different config structs. This change uses it for a command-line flag to try it. See next diff for more usage.

Reviewed By: dulinriley

Differential Revision: D17604269

fbshipit-source-id: ede0a3fa4c418adf58ffe0b2d263b82eb931818b