Record heap extents in custom breakpad data.
Summary:
I recently put memory protection on some portions of heap segments.  I realized that I might have a hard time determining whether a SEGV fault address was pointing into a heap segment.  This diff maintains lists of the segment addresses in custom breakpad data, on platforms that support it.

The diff takes care not do do quadratic work.  There is a breakpad key for every 10 segments; when we add a segment, we update only the key corresponding to the added segment.  When we do a full collection and remove segments, we remove any keys that no longer contain any segments.

Reviewed By: dulinriley

Differential Revision: D17661669

fbshipit-source-id: 58266b5d3dd2c9183bb4a1243a6df0f665cd3e91