Factor the input string tracking into a Cursor type
Summary:
The regex execution engine tracks its position in the input string with a
cursor pointer, with a start and end boundaries. However this traversal
logic will need to become more complicated with lookbehind assertions,
because the input string needs to be traversed backwards when in a
lookbehind.

In preparation for lookbehinds, introduce a new type Cursor which
encapsulates the bounds checking and traversal logic. This replaces uses
of raw pointers.

Reviewed By: avp

Differential Revision: D17608488

fbshipit-source-id: 502ffbcc74114698c29faba4670b16703e3125dd