Correctly handle unpaired surrogates in patterns
Summary:
In a Unicode regexp, if the pattern contains an unpaired surrogate, it can
only match against an unpaired surrogate from the input string. It must
not match against a paired surrogate which should be decoded.

Implement this behavior by always emitting literal surrogates as U16
instructions, forcing surrogates to be decoded from the input string.

Reviewed By: kodafb

Differential Revision: D17413818

fbshipit-source-id: 3797630a130da840c0a32bd8d413993917322884