Compare GCPointers directly in inline caches.
Summary:
This an optimization for compressed pointers.  Measurements indicated that decode operations (compressed --> 64-bit) outnumber encode (64-bit --> compressed) by more than 10 to 1.  In the decodes, for the BigGC synth benchmark, something like 1/6 of the decodes are for inline cache HiddenClass comparisons.  In the case of a cache hit, these decodes are unnecessary: what's important is that the cache key is a unique identifier of the cache, and that it be easily available.  The compressed form of a GCPointer fits this description.

Thus, this diff changes inline caches to use GCPointer<HiddenClass> as the key, rather than HiddenClass*.

Reviewed By: dulinriley

Differential Revision: D16751297

fbshipit-source-id: 4b183c999bec412dc13b6f77e6aa11b124a53308