Small opt to avoid a decode.
Summary: In measurements of where compressed pointer decodes were happening, I found a place where a CSE could save a significant number.  This diff does that CSE.  This is a tiny bit unsafe, since it lengthens the lifetime of a raw pointer over a call, but the call does not take a Runtime, and therefore cannot allocate.

Reviewed By: tmikov, kodafb

Differential Revision: D16751458

fbshipit-source-id: 78f3ae109f83719d6a4271976eab2d29abe3a633