Distinguish between implicit and explicit async breaks.
Summary:
When attempting to step over certain statements such as `console.log`,
the inspector can attempt to stop the debugger in order to print things,
inspect variables, etc.

This breaks an invariant in the debugger that we do not clear temporary
stepping breakpoints unless we have completed the step.

We cannot avoid this issue by ignoring async break requests during stepping,
because the user could explicitly request a break in an infinite loop
during a step over/out, for example.

To deal with this, we require API users to specify whether a pause is
implicit or explicit, and we pass that information through from the inspector.
We then have two bits in the async atomic, and we let the debugger know
on pause which kind of pause was requested.

If the pause was implicit, then we ignore it if we are currently stepping
because we'll likely pause soon.
To handle the infinite loop case, if the pause was explicit, we
clear any step state and do a hard pause as if the current point was a
breakpoint. Because any step state has been cleared, we're good to
perform more debugger actions as requested by the user.

Reviewed By: willholen

Differential Revision: D17857679

fbshipit-source-id: aff69c79e8ef3d9efaf1f1c2d25bdc938734a766