Fix step assert when a jump points to its next instruction
Summary:
When stepping, we set breakpoints at both the next instruction and to the jump
target. When the two locations were the same, we triggered an assert for
duplicated breakpoints.

This diff simply skips setting a breakpoint on the jump target if it
happens to be the same as the next instruction.

Reviewed By: haozhun

Differential Revision: D18103863

fbshipit-source-id: edb7636f76d7b104cc7e31f2f54ce91e9a12d3cd