Avoid relying on multiple iterations of segments being the same.
Summary:
The order in which modules were iterated when parsing the metadata file
and generating the SegmentRange were different than the order in which
they were iterated to call `generateIRForCJSModule`. The iteration of the
JSON file was performed after the JSON parser sorted the segments.

To avoid any inconsistencies here, we attach the module ID to the file
buffer and pass it through when calling `generateIRForCJSModule` and
populating `Module::cjsModules_`. This removes the undocumented assumptions.

Reviewed By: dulinriley

Differential Revision: D18067638

fbshipit-source-id: 939f19daa2c46ed40999e4ffe747597bc7a09165