Disable date-locale.js when UBSAN is on
Summary:
There's a UBSAN error in `icu/unifiedcache.h` (signed integer overflow of a hash code).
Can't modify third party code to either fix it or add a suppression, and I can't use a sanitize-blacklist
because the problem is in a header file which is included in a lot of cpp files.

Reviewed By: avp

Differential Revision: D18120270

fbshipit-source-id: 1cce3b2f5befa38241c0640670fba3688f5fbd79