Add summarization of old-gen vtables to detect corrupting writes.
Summary:
The crashes in the task are either some form of memory corruption between GCs, or some bug in full GC itself.  Memory corruption could either be something making a pointer bad, so that it does not point to an object head, or something is overwriting object headers (vtables) with garbage values.  (This is where I would bet.)

It would be nice to confirm (or deny) the memory corruption hypothesis.  This diff adapts a mechanism we used in a previous similar situation.  At the end of a full GC, traverse the live objects, adding up hashes of the vtable pointers of all the objects.  (And record the limits of the live objects.)  At the start of the next full GC, do this same computation.  If the results differ, then something has corrupted a vtable pointer.  We record this in Android breakpad custom data -- so when the bug occurs, we can tell if this corruption occurred.

Reviewed By: dulinriley, kodafb

Differential Revision: D18138538

fbshipit-source-id: 1217163999074d58513f9f11701aabf7c9ae9b4e