Add CopyableBasicString
Summary:
We rely on std::basic_string<> above certain length being byte-copyable.
Even though that is not guaranteed by any specification, any reasonable
non-debug implementation would naturally satisfy that requirement (we
also have a test that tries to detect it).

It was however discovered that the Microsoft STL "checked iterators"
feature prevents byte copy of std::basic_string<> because each container
gets a separately heap allocated proxy pointing back to the container.
Thus if the container is byte-copied, the proxy points to the old
location, which is now garbage.

To deal with this, we add a new type `CopyableBasicString`, which
depending on the implementation is either an alias of
`std::basic_string<>`, or a re-implementation of the std::basic_string
interface, which is however always safe to memcpy().

The reimplementation is actually very simple - it allocates a separate
std::basic_string on the heap, and contains a single pointer to it,
which is naturally safe to memcpy().

For now we enable one or another based on the value of Microsoft's macro
`_ITERATOR_DEBUG_LEVEL`. In the future we may add a build mode.

Reviewed By: kodafb

Differential Revision: D18132326

fbshipit-source-id: 9acd31f7f32bd883bff218c834c7450cf424fc27