Fix parsing of tagged template literal with supplementary chars
Summary:
When a slash is follwed by a supplementary-planes character in a tagged template
literal, the character is encoded as standard UTF-8 (using 4 code units)
instead of modified UTF-8 (where it is first broken into surrogate pairs,
and then UTF-8 encoded) in the special `raw` property.

In some cases, this leads to assertion failure in the compiler (and crash
in the runtime if the byte code is generated with assertion off).

Reviewed By: avp

Differential Revision: D18090003

fbshipit-source-id: 5f2f692ca348e2ca1acd24af9163298edbce4d76