Add targeted verification code for marking bug.
Summary: In the crashes that are the subject of the given task, we've determined that the problem (at least in some cases) is failure to follow a pointer in a YG collection -- so that the pointer is discovered to be to a bad address in full GC.  In these cases, we don't know, at the point of the crash, what referring object or root caused the bad mark bit to be marked.  This diff adds verification: when we follow a pointer to mark the referent, we verify that the referent is the head of a valid object.  The hermes_fatal should give us a stack trace that reveals the source of the reference.

Reviewed By: haozhun

Differential Revision: D18241537

fbshipit-source-id: 7f4fdcfb45a8f6008a3e1bacf60d262d2618d9e1