Always write out the synth trace in a pretty format
Summary:
There's no real need to conserve disk space when using traces, as it's only used
in debugging/local scenarios. The niceness of opening the trace file with vim is much
appreciated over the compactness of the file.
So turn on `pretty` mode by default whenever writing out a trace file.

Reviewed By: davedets

Differential Revision: D18234636

fbshipit-source-id: ccb501eb6137930ffc2c52e9ec4b9fc1de92ac02