Throw an exception for using un-traced values upon replay
Summary:
A common error that can happen when taking a synth trace is to forget to record environment
interactions.
The result of this is a `mathRandomSeed` set to 0, and empty arrays of calls to date functions.

Error early during replay with a more specific message, to make this problem obvious, rather than
fail at some indeterminate place in the future.

This exception won't be thrown for scripts that don't use `Date` or `Math.random` at all, which avoids having
to supply dummy values for such scripts (which tend to be small tests).

Reviewed By: davedets

Differential Revision: D18234635

fbshipit-source-id: 667e2afa44e3afd4d1fee338097316c6e9e2db6f