Print histogram of various array type sizes at the end of execution
Summary:
In order to help answer questions about the sizes of various arrays, do a heap pass at the end of execution
and print out the capacities and sizes of various arrays: `ArrayStorage`, `SegmentedArray`, `Segment`, and `JSArray`.
It will output the array's capacity, the number of instances with this capacity, and the percent of bytes
taken by this array capacity category.

This will only run in debug mode, and only when `-gc-print-stats` is on.

Reviewed By: davedets

Differential Revision: D17816464

fbshipit-source-id: dc96da05925996e230147ee9bf66a529fd7e638e