Add support for named internal properties
Summary:
Adds a way to allocate an internal property index by name in `InternalProperties.def`.

This mechanism is intended for properties that can be dynamically attached to objects of any type - and for which, therefore, we can't safely reserve an index in the anonymous internal property range, since different `JSObject` subclasses use the same anonymous property indices to mean different things.

NOTE: Named internal properties can only be added with `JSObject::addOwnProperty`, not `addInternalProperty`. Here we update the assertion in `addInternalProperty` to enforce this.

Reviewed By: tmikov

Differential Revision: D17526027

fbshipit-source-id: 19486023b685419577d6ba1cb70add63470565e4