Support multiple bytecode files in TraceInterpreter
Summary:
Sometimes multiple bytecode files are processed by Hermes in one execution, and this previously
would break synth tracing because it assumed only one file would be used.
Extend the system to allow multiple `BeginExecJS` events on different files.

File identities are tracked via a `sourceHash` which is a SHA-1 hash of the source file,
either computed from the source or extracted from the bytecode (where the compiler embeds the source
hash).

When the trace is taken, the sourceHash is recorded for each file that was executed. Upon replay, a set of
files with the same set of source hashes are expected. The source hash is used as a mapping key in the
implementation to associate a file given on a command line to a specific set of `BeginExecJS` calls.

As an escape hatch for the common case and backwards compatibility, if the sourceHash isn't present in a
BeginExecJS record (or is all zero), that is assumed to match a single file.

NOTE: `std::map` was used because there is no `std::hash` implementation provided for `std::array`.

Reviewed By: kodafb

Differential Revision: D18088907

fbshipit-source-id: 945e1c509eada5bae7a6e017ed67d82e5fbde079