Add support for locations of objects in heap snapshots
Summary:
Add support for locations to Hermes's heap snapshot format.
This means that an object can be annotated with a static source code location, so that
when a heap snapshot is taken while a debugger is on, you can jump directly to the source
line.

Future commits will have some objects in Hermes utilize this functionality.

Reviewed By: amnn

Differential Revision: D18498979

fbshipit-source-id: 9fbc32913a10c68cdecb6fe61ddcb2e6ed1ca8ff