Add HiddenClass native memory into heap snapshots
Summary:
`HiddenClass` had some native memory stored in its `WeakValueMap`, but that wasn't
being reported to heap snapshots at all.
Add some overloads of `snapshotAddNodes` and `snapshotAddEdges` to implement this.

The finalizer needs to be modified to make sure the native ID is untracked, else the pointer
could be re-used for something else accidentally.

Reviewed By: davedets

Differential Revision: D18466961

fbshipit-source-id: 6d05f6fb29c7bd5c5ae309ac55afdba2e2340a3c