Ensure that all native memory tracked by snapshots is cleaned up
Summary:
When native memory is tracked by the heap snapshot mechanism, it needs to associate an ID with a pointer.
If that pointer is free'd, it's important to remove the ID association so that in case the address is re-used (for example,
if `malloc` returns the pointer again) it isn't associated with the old ID.
The easiest way to check this is to make sure there aren't any native IDs left in the map once the runtime is being shut
down and all finalizers were run.

Reviewed By: davedets

Differential Revision: D18467960

fbshipit-source-id: dd1a15ee79ae080d81e420724e5628e657447de9