Fix problem with totalAllocatedBytes_ check.
Summary:
Previous diff [D18311229](D18311229) addressed [T27227998](https://our.intern.facebook.com/intern/tasks/?t=27227998): it added a redundant check to ensure that we were calculating the total allocated bytes correctly.  (We do something mildly clever to make the cost for this low in opt, so we check with a per-allocation counter.  The check is at destruction of the Runtime object.)
Turns out that this causes HaaS tests to fail in mode/dbg (where this redundant check is done.)  In fbcode, we treat out-of-memory as a C++ exception.  We have a HaaS test for this OOM behavior.   [D18311229](D18311229) incremented the counter before allocation -- this means the count is wrong in case the allocation fails with an exception.  So the check in the Runtime dtor finds the two allocated bytes counts disagree.

This diff fixes this.  It does it in the most straightforward (and arguably ugly) way: increment the debug counter only after an allocation has succeeded, just before returning the result.  You might think that an RAII class could be helpful here -- maybe we could do the increment in the dtor of such a class?  I went down that route, before I realized that of course the dtor would be executed in case of an OOM exception, and the original problem still exists.  But I mention this to point out the problem, in case someone can think of a better way to do this.

Reviewed By: dulinriley

Differential Revision: D18617632

fbshipit-source-id: bb242d0e1a1624409ed99ba6d64165df0a7b2b66