runtime flags to suppress HermesInternal
Summary:
HermesInternal is a potential attack surface and we want to disable as
much of it as possible. To this end introduce two new flags:

- `-enable-hermes-internal=false` hides all members of HermesInternal.
- `-Xhermes-internal-test-methods=bool` hides members needed only for
  testing, specifically `detachArrayBuffer()`.

Note that using the first flag causes a crash in this commit, because
the rest of Hermes is not ready for an empty HermesInternal. This will
be rectified in the subsequent commit.

Update affected tests and add a new test for
`-Xhermes-internal-test-methods`.

Reviewed By: avp

Differential Revision: D18651941

fbshipit-source-id: d9d170b29a580625f55fafb62cc6682b113cdd7e