Refactor getOwnPropertyNames/Symbols into getOwnPropertyKeys
Summary:
getOwnPropertyKeys now handles both Strings and Symbols, and
takes flags to specify which should be returned.  This is called by
the old functions, and also used to implement CopyDataProperties in
the Proxy case, and for Reflect.ownKeys.

This does add some extra tests to the code paths, but it's already a
lot of stuff, so I'm not worried about perf impact here.

Reviewed By: tmikov

Differential Revision: D18284964

fbshipit-source-id: 569edbbca0f13d8f69adcf91d23bf2406d24d6e6