Add overload of getComputedPropertyValue_RJS which takes the property name
Summary:
In order to distinguish non-existent properties from
properties which have undefined value, there is a common pattern which
uses getComputedPrimitiveDescriptor followed by
getComputedPropertyValue_RJS.  This pattern is inadequate in two ways.
First, in the places it's used, the spec generally specifies using the
abstract operation HasProperty, which will call the has trap on a
proxy, which would be incorrect to call from
getComputedPrimitiveDescriptor.  Second, there's no mechanism to call
the proxy get trap on a proxy.  This adds an overload for
getComputedPropertyValue_RJS which provides the proper abstraction to
solve both these problems.  The actual proxy functionality is added
later in the stack.

Reviewed By: avp

Differential Revision: D18284939

fbshipit-source-id: 196c57d5628388373c0f5481d68c024c4758b919