Implement nullish coalescing.
Summary:
Add the `??` operator, which is the third logical operator in JS,
as specified in the Stage 3 proposal:
https://tc39.es/proposal-nullish-coalescing/

It is defined as having lower precedence than `&&` or `||`.
However, because that is potentially confusing for the developer,
the grammar does not allow `??` in the same `ShortCircuitExpression`
as `&&` or `||`. We accommodate this requirement in `newBinNode`,
by keeping a couple flags across invocations and then ensuring
that we only allow `??` if neither `&&` or `||` have been used,
and vice versa. In practice, this should have minimal impact on
the performance of `parseBinaryExpression`.

IRGen for `??` is straightforward, as it follows the same
pattern as `&&` and `||`. The function was reworked slightly in order
to allow for three operators, and some confusing temporary variables
were eliminated for easier readability.

Reviewed By: dulinriley

Differential Revision: D18359522

fbshipit-source-id: 6348e3c9845d9dfa5c4bb85544391b41797e964a