Add optional UTF-16 surrogate pairs conversion in JSONParser
Summary:
`JSONParser` would store a string that had surrogate pairs as UTF-8 with
unconverted surrogates. This meant that using the characters as UTF-8 was mostly fine,
until a surrogate pair was found.

The fix for this is to add a mode flag to have JSLexer fix the encoding of these, by converting to UTF-16 and
back to UTF-8, fixing the surrogates while that happens.
This mode flag defaults to `false`, and only synth traces care about it right now.
This change should only minimally affect the normal parser performance.

Reviewed By: tmikov

Differential Revision: D18722155

fbshipit-source-id: a9fc140d77d223a872310efae7edadeb79af3711