Fix IRGen for optional chain deletion.
Summary:
The code `delete a?.b` is supposed to work, but the code neglected it.

This fix adds a `MemberExpressionOperation` argument to the `genMemberExpression`
functions which allows us to specify what to do with the member expression.

Added the functionality to the `genMemberExpression` call as well,
to funnel that IRGen through the same function for simplicity.

Reviewed By: tmikov

Differential Revision: D18813350

fbshipit-source-id: e545aee2a872ea92d5b7113ef61114fa75f49bdb