Refactor internal properties support
Summary:
When creating objects like JSRegExp that use anonymous internal properties, instead of adding the anonymous internal properties on every single instance creation, keep in Runtime a small list of hidden classes with 0...4 slots reserved, and directly use the appropriate hidden class.

This should have a small perf advantage, but the real purpose of this change is to prepare for overlapping the fields of subclasses of JSObject with the direct property storage (and eventually making anonymous internal properties redundant).

Reviewed By: tmikov

Differential Revision: D18138888

fbshipit-source-id: 6e2df731f18b3f45e0a47a94d50ac62c90fbbbeb