Make resizeWithinCapacity allocation-less for SegmentedArray
Summary:
`SegmentedArray::resizeWithinCapacity` was implemented in the sense of "the spine won't be moved if you
call this function"; however, the intention of the function was that calling it would not perform any allocations.

Change the implementation to respect this, by changing `capacity()` to mean "the number of elements that can be inserted before an allocation is required".
Add a new function `totalCapacity()` (private to `SegmentedArray`) that tracks the number of elements that can fit into the spine.

Update `resizeWithinCapacity` to no longer access a `runtime` parameter, so that it can't be made to allocate in
the future. Add `NoAllocScope` to callers to be especially sure.

This change is done so that `resizeWithinCapacity` is no longer a dangerous API to call when trimming is on.

Reviewed By: tmikov

Differential Revision: D18585546

fbshipit-source-id: dce64e113e063607252e4a7ad252310f33be7b80