Add specialized instructions for iteration.
Summary:
In a move to make for-of loops and array destructuring faster, we add
special instructions which will allow us to step an iterator using less
bytecode as well as less time.

There are 3 instructions in the IR with 1-1 equivalents in the bytecode
format and which are documented in IR.md:
- `IteratorBegin`
- `IteratorNext`
- `IteratorClose`

In the majority of cases, array patterns on the LHS come from a source
which is also an array, so the instructions allow skipping actual
construction of an Array iterator by simply storing the current index as
the iterator.

Reviewed By: tmikov

Differential Revision: D18604932

fbshipit-source-id: dbbb7a93d84554cd147066b8e63c60e1d4513744