Implement WeakMap marking for MallocGC.
Summary:
Extend previous work to scan weak maps specially to malloc GC, factoring out common parts into GCBase, and hardening against some subtle bugs.

A bug was that scanning a WeakMap object could conceivably cause a key of another WeakMap to become reachable (perhaps the first WeakMap has that key as a property).  We weren't handling that properly; the new implementation does.  As part of that, we have to introduce a new wrapping Acceptor that skips WeakRef marking -- we need to access the weak ref slots during WeakMap marking, so we postpone marking of WeakRefs in WeakMaps until after the WeakMap marking process is complete.

A complication here was that MallocGC's HANDLESAN mode is a form of copying collector.  For an in-place marking, the acceptor is idempotent: it is harmless to apply it to the same slot several times.  The copying collector acceptor is not normally idempotent however.  We introduced a mode to make it so, and use it in the portion of marking that may remark slots.

Reviewed By: dulinriley

Differential Revision: D18910521

fbshipit-source-id: de45e09932796160e17989d1cf632956b53b4461