Temporarily back out WeakMap fix.
Summary:
We've discovered a problem with the modification to WeakMap marking embodied by D18651375, D18758654, and D18910521.  So far it only occurs in production, and we haven't diagnosed.  This diff backs out the change.

I'm doing a simple "forward" change rather than a true backout.  Here's the argument for why this is safe:

The new code, in GCBase::completeWeakMapMarking in GCBase-inline.h, does nothing if the reachableWeakMaps is empty.  This change prevents anything from being added to that vector, reverting to pushing WeakMaps on the normal marking stack, as we did before the first diff.  I've been using this change to quickly disable the new behavior in performance testing, and have gotten the same GC behavior as before.

Reviewed By: tmikov

Differential Revision: D19190300

fbshipit-source-id: 9538dc71bbdc136e31f76203975b96f561f6561f