Fix totalTime for multi-bundle executions.
Summary: We discovered that the "totalTime" metric, and others, are wrong for executions that use multiple bundles: the "start" timer is reset at the start of each bytecode file that's executed.  So the total time is only the time from the last such bytecode file exec to the end, rather than from the first.  This diff fixes that.

Reviewed By: dulinriley

Differential Revision: D19181537

fbshipit-source-id: d9100b68f9773fabc297bd8515dc5a04d41e0998