Move scriptID to the RuntimeModule itself
Summary:
The script ID for a `RuntimeModule` was set and maintained by the debugger before,
but other callers might want to access this information.
Instead make it a field on `RuntimeModule`.

Note that in the case of lazy compilation, the script ID is only set on the root module, so make sure
to always access the script ID from there first.

Reviewed By: willholen

Differential Revision: D18740080

fbshipit-source-id: 8228bd1a780acb9352bd5a2a672c33b0d7d252c1