Compute overlapping direct property slots and shift other properties accordingly
Summary:
NOTE: This first diff does *not* move the direct property slots to always be trailing in the subclasses, hence does *not* actually overlap anything.

This diff introduces the `numOverlapSlots<Derived>()` function, which returns the number of direct property slots that will be unused in the class `Derived`. All other properties are shifted "up" by this amount, which requires changing the hardcoded indices used to access anonymous properties (`JSDate`, primitive boxes, `JSRegExp`, etc.) and the `.length` property of `JSArray`.

As a C++ technicality, the property indices must now be constexpr functions rather than constants, since they indirectly depend on `sizeof` of the class.

Reviewed By: tmikov

Differential Revision: D18838003

fbshipit-source-id: 2c474f4ecfb6154f366e70bbaefb9986ab1e28ec