Remove unnecessary config copy for StackRuntime
Summary:
The `GCConfig` was being copied for `StackRuntime`, but it no longer needed that config
(it went back to using `mmapProvider` a while ago).
This copy was problematic because it ran a destructor on the `runtimeMemoryThread`.
One of the fields of the config was a `std::function` that happened to hold a `jni::global_ref`.
Those references can't be destructed on a thread not registered with the JVM, which
`runtimeMemoryThread` wasn't.

Reviewed By: mhorowitz

Differential Revision: D19277414

fbshipit-source-id: 0c12c9590132fee1ba9a841a97d656da6b1b5b06