Add number pseudo-nodes into heap snapshots
Summary:
An unfortunate aspect of Chrome heap snapshots is that number values
are not displayed if they aren't allocated on the heap. For Hermes this means
all numbers are not displayed (for example, `.length` on an array cannot be inspected).

There's a workaround for this problem: pretend that all numbers are singleton objects on
the heap, and all `HermesValue` with that numeric value is a pointer to that singleton.
Then add all those singletons as zero-size nodes in the snapshot, and all the "pointers" to it as an edge.

As a consequence of these being zero-size, they are not added to the heap size totals, but they are
also not displayed in the snapshot in their own `(number)` section. So "Reveal in Summary" doesn't
work for them. That's better than the alternative, which is to skew the heap size numbers though.

Add this functionality to all internal values, `JSObject` properties, and `JSArray` elements.

Reviewed By: davedets

Differential Revision: D19306601

fbshipit-source-id: a207dba7e04ed7382fa0986058bdc6aea893cdcc