Treat global function less special in debug lexical info
Summary:
Treat global function less special in parsing of debug lexical info (in debugger) and generation of debug lexical info (in ISel).

In a later diff, environment register is going to be introduced into the lexical debug info. This field is meaningful and necessary to the global function. Previously, the global function is excluded from having debug info. This diff removes the exclusion to pave the way for the upcoming diff.

Reviewed By: avp

Differential Revision: D17784264

fbshipit-source-id: da89f17f93b52de18d1f6df90b55832417b9e55e