Track host function name and number of parameters
Summary:
`HostFunction` should also track its name and number of parameters at the time of its
creation, just in case something relies on that behavior in the wild.

The trace format is backwards-compatible, and will still parse and exec existing traces.

The default name if none is specified is the empty string, which will be reported as `"anonymous"`
in various places automatically. The default `paramCount` is 0, which matches what `TraceInterpreter`
did before this change.

Reviewed By: davedets

Differential Revision: D19221486

fbshipit-source-id: 283b6e0a72d9d017f3e4957cb038a7edfc303ee1