Minscript POC
Summary:
Minscript, or `mins` for short, is a minimalistic bytecode interpreter
for scripting. This is the 3rd proof of concept version and also the
simplest. Previous versions experimented with other approaches and more
complexity.

The bytecode is stack based and is designed to support the following
features:
- lexical closures
- classes with static keys and inheritance
- arrays with integer keys
- dicts with dynamic keys
- strings, floating point numbers, bools, undefined
- the usual control flow constructs
- exceptions with try/catch/finally

The interpreter itself is intended to support:
- optional garbage collection (+1KB binary size)
- bump pointer allocation from a pre-existing buffer
- very fast startup
- ffi

At this time enough has been implemented in order to run the
"interpreter-dispatch" benchmark, which calculates factorial in a loop.

The performance in that benchmark in the POC is about 3x slower than
Hermes. The size of a standalone Android ARM32 executable is 17K, of
which 10KB is actual code, so the expected impact when linked in an
existing binary is about that.

It builds in the Hermes directory for convenience, but the intent is for
the runtime to have no dependencies on Hermes or LLVM (the latter is too
much hassle for a minimalistic interpreter).

Reviewed By: avp

Differential Revision: D18568343

fbshipit-source-id: df7b4cb3f2daeb94b3ef708f85af84a7de54c4fb