Add more detailed function meta-data output
Summary:
Renames the `offset` command to `function-info` and expands it to include information on: function name, the actual source location, and the origin source location. The latter requires a source-map to be provided on the command line.

Using this command it's now possible to find out where all the functions are in a HBC file and how they map back to the original source.

Reviewed By: kodafb

Differential Revision: D19213728

fbshipit-source-id: 0e6898fb19a1171b82f2a68794a42ba88b6edeb7