Add trace/replay of HermesInternal.getInstrumentedStats()
Summary:
Synth trace replay started crashing recently.  This turned out to be due to a source of non-determinism we had noticed previously.  See the test plan for why.

This diff treats the results of HermesInternal.getInstrumentedStats() as we do, for example, Date.now(): we log the results during the original execution, save these logs as part of the trace, and substitute the logged results for the current results on replay.

Reviewed By: dulinriley

Differential Revision: D19844732

fbshipit-source-id: ecd43d3594dbb3ebab4a501ceea022dfd600c927