Improve build and add instructions
Summary:
Tweak the CMake build to better support Emacripten and add
instructions.

## Performance

When built in Release+Wasm mode and executed by NodeJS, the Hermes
compiler itself is about 5-10% slower than a native version when compiling
a multi-megabyte JS source. It is likely that tweaking the memory growth
parameters will help even further.

VM execution performance of interp-dispatch.js is
about 6 times slower with Wasm Hermes compared to a native one. I
haven't spent time to look into it though. It is likely that it can be
improved.

Reviewed By: dulinriley

Differential Revision: D19804100

fbshipit-source-id: d808fa9235d28b3217504e51ddd0d408befeb534