Ensure ScopedNativeCallFrame overflow is checked everywhere
Summary: ScopedNativeCallFrame requires the user to always check the overflow flag before accessing a newly constructed instance. This is dangerous if the user forgets, which would only show up on an actual overflow in debug mode. To protect against the user forgetting, update the asserts in the class to not only check that the overflow flag is false, but that is was checked to be false by the user.

Reviewed By: tmikov

Differential Revision: D20122661

fbshipit-source-id: b3e03f3aa3788c3ca3212b0c27b82f51dbce6d52