Make sure that we only override trace config values if requested.
Summary:
Experiments show that the config used during replay is *not* entirely the config recorded in the trace.  The "synth" program allows a large set of runtime flags, and those of these that are acted upon overwrite the trace's config with the flag's default values.  This is not good for replay determinism.  (Nor is it good, really, for benchmarking -- we really want to be running with similar settings, I think.)

This diff only overrides the config values from the trace if the corresponding command line flag is explicitly specified.  After searching a bit, I figured out that llvm::cl::opt::getNumOccurrences() allows you determine this.  The fields in TraceInterpreter::ExecuteOptions that correspond to config parameters becomes LLVM::Optionals, with default value None.  We only override the config parameter if the ExecuteOptions field has a value.

Reviewed By: dulinriley

Differential Revision: D19910876

fbshipit-source-id: f95cede57bf85b5fd3e65f39c32655a34d854934