Add feature to track JS stack-trace for JS heap allocated objects
Summary:
This diff adds the minimal infrastructure needed to recover a stack-trace for each JS heap allocated object. At its core is a shared tree data-structure representing all stack-traces. This data-structure is held by `Runtime`, and tracks where we are in the current call-stack with push + pop calls integrated into the interpreter loop. A new `AllocLocTracker` utility class in `GCBase` is used by GC implementations to associate allocations with stack-traces and keep them updated.

As this feature adds run-time overhead in the interpreter it's only enabled with `HERMES_ENABLE_DEBUGGER`. This means a debug VM must be used for memory allocation profiling and so execution will be slower. However, data collected by a debug VM should still be fairly accurate compared to a production VM, provided bytecode is built with optimizations.

Reviewed By: dulinriley

Differential Revision: D20383004

fbshipit-source-id: 52a406e9851bbbf43910865577a96da295cb8d0b