add a Wasm compiler driver to use from JavaScript
Summary:
tools/emhermesc is a new target. When compiled with Emscripten, it
exports an interface for efficient compilation from a JS string to a
NodeJS Buffer containing bytecode.

The interface is very minimalistic for now, but it can be extended in
the future.

In the same directory there is a simpolistic JavaScript wrapper
illustrating the usage.

Reviewed By: cpojer

Differential Revision: D20811491

fbshipit-source-id: 650c179e118eac3481d873be2ad4bc8038e48a9f