Register every RuntimeModule and BCProviderFromBuffer with the CrashManager
Summary: Register every RuntimeModule and BCProviderFromBuffer with the CrashManager. Since Runtime is already registered, this makes it possible during debugging to iterate over the list of runtime modules and look at the location and size of their bytecode buffers. The *contents* of bytecode buffers are not included, since that might be too large.

Reviewed By: tmikov

Differential Revision: D20844755

fbshipit-source-id: 1e7fc1bdd1c0234e47591fb033d0cb0147882e3f