Fix TestHelpers createCodeBlock memory leak
Summary:
The `createCodeBlock` function from TestHelpers was used in a couple tests, but
didn't have a link between the `RuntimeModule` and the new `CodeBlock`.
This resulted in that `CodeBlock *` getting leaked.
The fix is simple: use `getCodeBlockMayAllocate` which properly initializes the
`RuntimeModule::functionMap_` entry, so that the destructor will free the code block memory.

Reviewed By: avp

Differential Revision: D20998412

fbshipit-source-id: f3375b16710f164721e0df947936a1680b16c26f