dynamic optimization callback
Summary:
Until now BytecodeProviderFromSrc only invoked the IR optimization
passes if `compileFlags.optimize == true`, and
HERMESVM_ENABLE_OPTIMIZATION_AT_RUNTIME is defined to true. The intent
was to avoid linking the optimization passes in the VM if we don't need
them at runtime.

There are two problems with this:
- The API behaves inconsistently. A caller might request optimizations,
but they will not be performed depending on a build flag which the
caller may not even be aware of.
- If the build flag is not set, it is impossible to perform
optimizations at all. To get reliable results one is forced to copy the
implementation of BytecodeProviderFromSrc.

To solve this, we simply add a new static factory function accepting a
callback to perform optimizations. This defers the decision of whether
to link the optimization passes to the caller. Then can pass `nullptr`
(or an empty function), or they can pass `hermes::runAllOptimzations`,
or something in between.

For now preserve the existing unreliable API (only now it calls the new
one), but the hope is to be able to remove it soon.

Reviewed By: kodafb

Differential Revision: D20811492

fbshipit-source-id: 72f125705fea2204fe04fa3982746b7901005ef4