Introduce DecoratedObject and rebuild HostObject on top of it (#225)
Summary:
This PR conceptually factors out the "holds a native pointer" aspect of
HostObject, allowing it to be used independently from the "interposes properties."
The new intermediate class is called DecoratedObject.

This fills an integration gap.  Say you wish to attach a native thing to a JS object,
and manipulate the native bits from prototype functions. This is not possible
with HostObject, because the prototype is not searched. DecoratedObject is
the missing intermediate layer: it manages the native bits (memory management,
etc) without affecting the object model.

This PR squashes the following commits:

## Switch HostObject to storing unique_ptr instead of shared_ptr

HostObject currently stores a shared_ptr, but in practice the refcount is
always 1; jsi does not share objects. Switch to storing unique_ptr instead.
This both reduces allocation size and prepares for introducing
DecoratedObject.

## Introduce DecoratedObject

Hermes currently has no easy way to associate a finalizable pointer with an
JSObject. This commit introduces a new JSObject subclass, called
DecoratedObject, to fill this gap.

A DecoratedObject holds a unique_ptr to a Decoration, which is intended to
be subclassed. There are ways of getting and setting the Decoration, and
also for the Decoration to report its memory usage.

The Decoration is purely for "integration" uses. JSObject does not care about
the decoration.

The next commit proposes rebuilding HostObject atop DecoratedObject, which
reduces some duplication.

## HostObject to subclass DecoratedObject

This "rebuilds" HostObject on top of DecoratedObject. HostObjectProxy now
subclasses DecoratedObject::Decoration. The advantage is to reduce code
duplication: HostObject can take advantage of the memory management
of DecoratedObject.

The third commit rebuilds HostObject on top of DecoratedObject; now HostObject only does the property-interposing parts.
Pull Request resolved: https://github.com/facebook/hermes/pull/225

Reviewed By: tmikov

Differential Revision: D21211233

Pulled By: mhorowitz

fbshipit-source-id: 2e3e1100f00ed765df85be19231744445c328f1c