Add bool to BeginExecJS record indicating whether input was bytecode.
Summary:
The BeginExecJSRRecord type in synth trace records the hash of the JS source, but it doesn't record whether the original execution was from source, or from the bytecode compiled from that source.  This diff adds a boolean property to the record to indicate that.

We will probably use this in the script that collects the source or bytecode from the build artifacts, so it can find the .hbc or .js files, appropriately.

Reviewed By: dulinriley

Differential Revision: D20818603

fbshipit-source-id: 5d05cdbd5786061c8d6b673c1619880c9ce34d39