Keep a record of GC executions.
Summary:
To determine whether we have precisely reproduced the sequence of GCs (and allocations) between the original execution and a trace, we need to keep a record of those GCs (and allocations) in the trace.  This diff does that.  Note that fine-grained allocations are only tracked prior to the first YG GC for now -- these would make the trace very large.

Currently, this is only used when we compile with a new hermes.api_trace_debug config parameter.  In the future, we may record the GC sequence always, and extend it to have the set of Object IDs reachable from external roots, so that we can only use the external roots observed during the original execution.

Reviewed By: dulinriley

Differential Revision: D20448892

fbshipit-source-id: 3ff945a07a281d760aea1afa0900d1f692264719